+++ tests/TestCase/Command/LinkScannerCommandTest.php
@@ -20,7 +20,6 @@ use Cake\Console\ConsoleOptionParser;
 use Cake\Console\Exception\StopException;
 use Cake\TestSuite\Stub\ConsoleOutput;
 use LinkScanner\Command\LinkScannerCommand;
-use LinkScanner\TestSuite\IntegrationTestTrait;
 use LinkScanner\TestSuite\TestCase;
 use LinkScanner\Utility\LinkScanner;
 use MeTools\TestSuite\ConsoleIntegrationTestTrait;
@@ -30,9 +29,7 @@ use MeTools\TestSuite\ConsoleIntegrationTestTrait;
  */
 class LinkScannerCommandTest extends TestCase
 {
-    use ConsoleIntegrationTestTrait, IntegrationTestTrait {
-        ConsoleIntegrationTestTrait::configApplication insteadof IntegrationTestTrait;
-    }
+    use ConsoleIntegrationTestTrait;
 
     /**
      * @var \LinkScanner\Utility\LinkScanner
@@ -217,7 +214,11 @@ class LinkScannerCommandTest extends TestCase
         }
 
         //Enables follow redirects
-        $this->LinkScanner = $this->getLinkScannerClientReturnsFromTests();
+        $this->LinkScanner = $this->getMockBuilder(LinkScanner::class)
+            ->setConstructorArgs([$this->getClientReturnsRedirect()])
+            ->setMethods(['_createLockFile'])
+            ->getMock();
+
         $this->Command->LinkScanner = $this->LinkScanner;
         $this->getEventManager();
         array_pop($params);
@@ -225,7 +226,8 @@ class LinkScannerCommandTest extends TestCase
         $this->assertEventFired('LinkScanner.foundRedirect', $this->LinkScanner->getEventManager());
         $this->assertContains(['followRedirects' => true], $this->LinkScanner->getConfig());
         $this->assertOutputContains('Redirects will be followed');
-        $this->assertOutputContains('Redirect found: http://localhost/pages/third_page');
+        $this->assertOutputContains('Redirect found: http://localhost/redirectTarget');
+        $this->assertOutputContains('Checking http://localhost/redirectTarget ...');
 
         //With an invalid full base url
         $this->expectException(StopException::class);